improve triangulation algorithm, cut into separate file
